move main-related tsx to src/main
