+++ index.html
@@ -10,6 +10,6 @@
   <body>
     <span style="display:none;" id="__csrftoken__" value="__CSRFTOKEN__"></span>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/main"></script>
   </body>
 </html>
