--- src/changeEmail.tsx
@@ -1,4 +1,0 @@
-import mainCore from "./mainCore";
-import Routes from "./routes/ChangeEmail";
-
-mainCore(Routes);
