+++ src/i18n/index.ts
@@ -1,5 +1,5 @@
 import i18n from "i18next";
-import LanguageDetector from "i18next-browser-languagedetector";
+// import LanguageDetector from "i18next-browser-languagedetector";
 import { initReactI18next } from "react-i18next";
 import en from "./en";
 import zhTW from "./zh-TW";
