--- src/login.tsx
@@ -1,4 +1,0 @@
-import mainCore from "./mainCore";
-import Routes from "./routes/Login";
-
-mainCore(Routes);
