+++ src/main/all.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/All";
+import core from "./core";
+
+core(Routes);
