+++ src/main/attemptChangeEmail.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/AttemptChangeEmail";
+import core from "./core";
+
+core(Routes);
