+++ src/main/attemptSetIDEmail.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/AttemptSetIDEmail";
+import core from "./core";
+
+core(Routes);
