+++ src/main/changeEmail.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/ChangeEmail";
+import core from "./core";
+
+core(Routes);
