+++ src/main/changePasswd.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/ChangePasswd";
+import core from "./core";
+
+core(Routes);
