+++ src/main/core.tsx
@@ -1,31 +1,31 @@
-import "./vendors";
-import "./index.css";
+import "../vendors";
+import "../index.css";
 import { registerThunk } from "@chhsiao1981/use-thunk";
 import config from "config";
 import { type FC, StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import reportWebVitals from "./reportWebVitals";
-import * as DoArticlePage from "./thunks/articlePage";
-import * as DoArticlesPage from "./thunks/articlesPage";
-import * as DoAttemptChangeEmailPage from "./thunks/attemptChangeEmailPage";
-import * as DoAttemptSetIDEmailPage from "./thunks/attemptSetIDEmailPage";
-import * as DoChangeEmailPage from "./thunks/changeEmailPage";
-import * as DoChangePasswdPage from "./thunks/changePasswdPage";
-import * as DoClassBoardsPage from "./thunks/classBoardsPage";
-import * as DoGeneralBoardsPage from "./thunks/generalBoardsPage";
-import * as DoHeader from "./thunks/header";
-import * as DoHomePage from "./thunks/homePage";
-import * as DoHotBoardsPage from "./thunks/hotBoardsPage";
-import * as DoInitPage from "./thunks/initPage";
-import * as DoLoginPage from "./thunks/loginPage";
-import * as DoManualPage from "./thunks/manualPage";
-import * as DoManualsPage from "./thunks/manualsPage";
-import * as DoNewArticlePage from "./thunks/newArticlePage";
-import * as DoProfilePage from "./thunks/profilePage";
-import * as DoRegisterPage from "./thunks/registerPage";
-import * as DoSetIDEmailPage from "./thunks/setIDEmailPage";
-import * as DoUserFavoritesPage from "./thunks/userFavoritesPage";
-import "./i18n";
+import reportWebVitals from "../reportWebVitals";
+import * as DoArticlePage from "../thunks/articlePage";
+import * as DoArticlesPage from "../thunks/articlesPage";
+import * as DoAttemptChangeEmailPage from "../thunks/attemptChangeEmailPage";
+import * as DoAttemptSetIDEmailPage from "../thunks/attemptSetIDEmailPage";
+import * as DoChangeEmailPage from "../thunks/changeEmailPage";
+import * as DoChangePasswdPage from "../thunks/changePasswdPage";
+import * as DoClassBoardsPage from "../thunks/classBoardsPage";
+import * as DoGeneralBoardsPage from "../thunks/generalBoardsPage";
+import * as DoHeader from "../thunks/header";
+import * as DoHomePage from "../thunks/homePage";
+import * as DoHotBoardsPage from "../thunks/hotBoardsPage";
+import * as DoInitPage from "../thunks/initPage";
+import * as DoLoginPage from "../thunks/loginPage";
+import * as DoManualPage from "../thunks/manualPage";
+import * as DoManualsPage from "../thunks/manualsPage";
+import * as DoNewArticlePage from "../thunks/newArticlePage";
+import * as DoProfilePage from "../thunks/profilePage";
+import * as DoRegisterPage from "../thunks/registerPage";
+import * as DoSetIDEmailPage from "../thunks/setIDEmailPage";
+import * as DoUserFavoritesPage from "../thunks/userFavoritesPage";
+import "../i18n";
 
 export default (Routes: FC) => {
   registerThunk(DoArticlePage);
