+++ src/main/dev.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/Dev";
+import core from "./core";
+
+core(Routes);
