+++ src/main/error.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/Error";
+import core from "./core";
+
+core(Routes);
