+++ src/main/home.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/Home";
+import core from "./core";
+
+core(Routes);
