+++ src/main/init.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/Init";
+import core from "./core";
+
+core(Routes);
