+++ src/main/login.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/Login";
+import core from "./core";
+
+core(Routes);
