+++ src/main/register.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/Register";
+import core from "./core";
+
+core(Routes);
