+++ src/main/setIDEmail.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/SetIDEmail";
+import core from "./core";
+
+core(Routes);
