+++ src/main/userInfo.tsx
@@ -1,0 +1,4 @@
+import Routes from "../routes/UserInfo";
+import core from "./core";
+
+core(Routes);
